+++ public/filesystem.h
@@ -420,7 +420,7 @@ public:
 	virtual bool			IsFileWritable( char const *pFileName, const char *pPathID = 0 ) = 0;
 	virtual bool			SetFileWritable( char const *pFileName, bool writable, const char *pPathID = 0 ) = 0;
 
-	virtual long			GetFileTime( const char *pFileName, const char *pPathID = 0 ) = 0;
+	virtual int32_t			GetFileTime( const char *pFileName, const char *pPathID = 0 ) = 0;
 
 	//--------------------------------------------------------
 	// Reads/writes files to utlbuffers. Use this for optimal read performance when doing open/read/close
@@ -505,7 +505,7 @@ public:
 	// File I/O and info
 	virtual bool			IsDirectory( const char *pFileName, const char *pathID = 0 ) = 0;
 
-	virtual void			FileTimeToString( char* pStrip, int maxCharsIncludingTerminator, long fileTime ) = 0;
+	virtual void			FileTimeToString( char* pStrip, int maxCharsIncludingTerminator, int32_t fileTime ) = 0;
 
 	//--------------------------------------------------------
 	// Open file operations
@@ -716,7 +716,7 @@ public:
 	virtual bool		FullPathToRelativePathEx( const char *pFullpath, const char *pPathId, char *pRelative, int maxlen ) = 0;
 
 	virtual int			GetPathIndex( const FileNameHandle_t &handle ) = 0;
-	virtual long		GetPathTime( const char *pPath, const char *pPathID ) = 0;
+	virtual int32_t		GetPathTime( const char *pPath, const char *pPathID ) = 0;
 
 	virtual DVDMode_t	GetDVDMode() = 0;
