+++ public/tier0/basetypes.h
@@ -151,17 +151,17 @@ typedef float vec_t;
 
 #ifdef __cplusplus
 
-inline unsigned long& FloatBits( vec_t& f )
+inline uint32_t& FloatBits( vec_t& f )
 {
-	return *reinterpret_cast<unsigned long*>(&f);
+	return *reinterpret_cast<uint32_t*>(&f);
 }
 
-inline unsigned long const& FloatBits( vec_t const& f )
+inline uint32_t const& FloatBits( vec_t const& f )
 {
-	return *reinterpret_cast<unsigned long const*>(&f);
+	return *reinterpret_cast<uint32_t const*>(&f);
 }
 
-inline vec_t BitsToFloat( unsigned long i )
+inline vec_t BitsToFloat( uint32_t i )
 {
 	return *reinterpret_cast<vec_t*>(&i);
 }
@@ -171,7 +171,7 @@ inline bool IsFinite( vec_t f )
 	return ((FloatBits(f) & 0x7F800000) != 0x7F800000);
 }
 
-inline unsigned long FloatAbsBits( vec_t f )
+inline uint32_t FloatAbsBits( vec_t f )
 {
 	return FloatBits(f) & 0x7FFFFFFF;
 }
@@ -215,7 +215,7 @@ inline float FloatNegate( vec_t f )
 }
 
 
-#define FLOAT32_NAN_BITS     (unsigned long)0x7FC00000	// not a number!
+#define FLOAT32_NAN_BITS     (uint32_t)0x7FC00000	// not a number!
 #define FLOAT32_NAN          BitsToFloat( FLOAT32_NAN_BITS )
 
 #define VEC_T_NAN FLOAT32_NAN
@@ -327,7 +327,7 @@ protected:
 
 
 template< class DummyType >
-class CIntHandle32 : public CBaseIntHandle< unsigned long >
+class CIntHandle32 : public CBaseIntHandle< uint32_t >
 {
 public:
 	inline			CIntHandle32() {}
