+++ public/tier1/utlmemory.h
@@ -661,7 +661,7 @@ inline bool CUtlMemory<T,I>::IsIdxValid( I i ) const
 {
 	// GCC warns if I is an unsigned type and we do a ">= 0" against it (since the comparison is always 0).
 	// We get the warning even if we cast inside the expression. It only goes away if we assign to another variable.
-	long x = i;
+	int32_t x = i;
 	return ( x >= 0 ) && ( x < m_nAllocationCount );
 }
